Minimize the number of rerenders by splitting sheet logic into separate component
